+++ airflow/contrib/hooks/sagemaker_hook.py
@@ -26,8 +26,8 @@ import warnings
 from botocore.exceptions import ClientError
 
 from airflow.contrib.hooks.aws_hook import AwsHook
-from airflow.contrib.hooks.aws_logs_hook import AwsLogsHook
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 from airflow.utils import timezone
