+++ airflow/contrib/operators/ecs_operator.py
@@ -22,9 +22,9 @@ from datetime import datetime
 from typing import Optional
 
 from airflow.contrib.hooks.aws_hook import AwsHook
-from airflow.contrib.hooks.aws_logs_hook import AwsLogsHook
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
+from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
 from airflow.typing_compat import Protocol
 from airflow.utils.decorators import apply_defaults
