+++ tests/contrib/sensors/test_sagemaker_training_sensor.py
@@ -22,10 +22,10 @@ from datetime import datetime
 
 import mock
 
-from airflow.contrib.hooks.aws_logs_hook import AwsLogsHook
 from airflow.contrib.hooks.sagemaker_hook import LogState, SageMakerHook
 from airflow.contrib.sensors.sagemaker_training_sensor import SageMakerTrainingSensor
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
 
 DESCRIBE_TRAINING_COMPELETED_RESPONSE = {
     'TrainingJobStatus': 'Completed',
